fix: support video files larger than max int size in bytes

+++ src/Skybrud.VideoPicker/Models/Videos/VideoFile.cs
@@ -34,7 +34,7 @@ namespace Skybrud.VideoPicker.Models.Videos {
             Height = obj.GetInt32("height");
             Url = obj.GetString("url");
             Type = obj.GetString("type");
-            Size = obj.GetInt32("size");
+            Size = obj.GetInt64("size");
         }
 
         public static VideoFile Parse(JObject obj) {
